Modal Add & modal edit

+++ redux-app/src/components/Table.js
@@ -3,25 +3,23 @@ import { deleteAPI, getAPI } from '../actions/Action';
 import { connect } from 'react-redux';
 import { Button, ButtonToolbar } from 'react-bootstrap';
 import ModalEdit from './ModalEdit';
+import ModalAdd from './ModalAdd';
 
 class Table extends Component {
     constructor(props) {
         super(props);
-        this.state = { props: [], openModal: false };
+        this.state = { openModalEdit: false, openModalAdd: false };
     }
 
     componentDidMount() {
         this.loadList();
     }
 
-    edit(model) {
-
-        this.setState({ model: model });
-
-        //this.props.changeModel(model);
-    }
-
     delete(id) {
+        if (!window.confirm("Delete?")) {
+            return;
+        }
+
         deleteAPI(id).then(resp => {
             if (resp.ok) {
                 this.loadList();
@@ -35,14 +33,15 @@ class Table extends Component {
 
     render() {
         const { descricao, id, valor, quantidade } = this.state;
-        let closeModal = () => this.setState({ openModal: false });
+        let closeModalAdd = () => this.setState({ openModalAdd: false });
+        let closeModalEdit = () => this.setState({ openModalEdit: false });
         return (
             <div>
                 <ButtonToolbar>
-                    <Button onClick={() => this.setState({ model: {}, openModal: true })}>
+                    <Button onClick={() => this.setState({ model: {}, openModalAdd: true })}>
                         Add
                     </Button>
-                    <ModalEdit model={this.state.model} show={this.state.openModal} onHide={closeModal} />
+                    <ModalAdd show={this.state.openModalAdd} onHide={closeModalAdd} />
                 </ButtonToolbar>
                 <table className="table table-striped table-dark">
                     <thead>
@@ -51,7 +50,7 @@ class Table extends Component {
                             <th scope="col">Description</th>
                             <th scope="col">Quantity</th>
                             <th scope="col">Price</th>
-                            <th scope="col">Action</th>
+                            <th scope="col">Actions</th>
                         </tr>
                     </thead>
                     <tbody>
@@ -63,17 +62,14 @@ class Table extends Component {
                                     <td>{item.quantidade}</td>
                                     <td>{item.valor}</td>
                                     <td>
-                                    <ButtonToolbar>
-                                        <Button onClick={() => this.setState({ descricao: item.descricao, valor: item.valor, id: item.id, quantidade: item.quantidade, openModal: true })}>
-                                            Edit
+                                        <ButtonToolbar>
+                                            <Button variant="success" onClick={() => this.setState({ descricao: item.descricao, valor: item.valor, id: item.id, quantidade: item.quantidade, openModal: true })}>
+                                                Edit
                                         </Button>
-                                        <ModalEdit descricao={descricao} id={id} valor={valor} quantidade={quantidade} show={this.state.openModal} onHide={closeModal} />
-                                    </ButtonToolbar>
-                                        {/* <div className="btn-group" role="group" aria-label="Basic example">
-                                            <button type="button" className="btn btn-success" onClick={() => this.setState({ descricao: item.descricao, openModal: true })} >Edit</button>
-                                            <button type="button" className="btn btn-success" onClick={() => { this.edit(item) }}>Edit</button>
-                                            <button type="button" className="btn btn-danger" onClick={() => { this.delete(item.id) }}>Delete</button>
-                                        </div> */}
+                                            <Button variant="danger" onClick={() => { this.delete(item.id) }}>Delete</Button>
+
+                                            <ModalEdit descricao={descricao} id={id} valor={valor} quantidade={quantidade} show={this.state.openModalEdit} onHide={closeModalEdit} />
+                                        </ButtonToolbar>
                                     </td>
                                 </tr>)
                         }
@@ -84,21 +80,8 @@ class Table extends Component {
     }
 }
 
-const mapStateToProps = (state) => {
-    return {
-        model: state.model
-    }
-}
-
 const mapDispatchToProps = (dispatch) => {
     return {
-        changeModel: (model) => {
-            dispatch({
-                type: "CHANGE_MODEL",
-                payload: model,
-                key: 'model'
-            })
-        },
         changeList: (list) => {
             dispatch({
                 type: "CHANGE_LIST",
@@ -109,4 +92,4 @@ const mapDispatchToProps = (dispatch) => {
     }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Table);
+export default connect(null, mapDispatchToProps)(Table);
